requirements.txt
google-auth-oauthlib>=1.0.0
google-api-python-client>=2.0.0
requests>=2.31.0
python-dotenv>=1.0.0
pydantic>=2.5.0
pytest>=7.4.0
aiohttp>=3.9.0
cachetools>=5.3.0